+++ src/pretix/plugins/badges/views.py
@@ -1,13 +1,25 @@
+import json
+from io import BytesIO
+
 from django.contrib import messages
+from django.contrib.staticfiles import finders
+from django.core.files import File
+from django.core.files.storage import default_storage
 from django.db import transaction
 from django.http import Http404
 from django.shortcuts import redirect
+from django.templatetags.static import static
 from django.urls import reverse
+from django.utils.functional import cached_property
 from django.utils.translation import ugettext_lazy as _
-from django.views import View
 from django.views.generic import CreateView, DeleteView, DetailView, ListView
+from reportlab.lib import pagesizes
+from reportlab.pdfgen import canvas
 
+from pretix.base.models import CachedFile, OrderPosition
+from pretix.base.pdf import Renderer
 from pretix.control.permissions import EventPermissionRequiredMixin
+from pretix.control.views.pdf import BaseEditorView
 from pretix.plugins.badges.forms import BadgeLayoutForm
 
 from .models import BadgeLayout
@@ -29,12 +41,7 @@ class LayoutCreate(EventPermissionRequiredMixin, CreateView):
     template_name = 'pretixplugins/badges/edit.html'
     permission = 'can_change_event_settings'
     context_object_name = 'layout'
-
-    def get_success_url(self) -> str:
-        return reverse('plugins:badges:index', kwargs={
-            'organizer': self.request.event.organizer.slug,
-            'event': self.request.event.slug,
-        })
+    success_url = '/ignored'
 
     @transaction.atomic
     def form_valid(self, form):
@@ -42,9 +49,14 @@ class LayoutCreate(EventPermissionRequiredMixin, CreateView):
         if not self.request.event.badge_layouts.filter(default=True).exists():
             form.instance.default = True
         messages.success(self.request, _('The new badge layout has been created.'))
-        ret = super().form_valid(form)
-        form.instance.log_action('pretix.plugins.badges.layout.added', user=self.request.user, data=dict(form.cleaned_data))
-        return ret
+        super().form_valid(form)
+        form.instance.log_action('pretix.plugins.badges.layout.added', user=self.request.user,
+                                 data=dict(form.cleaned_data))
+        return redirect(reverse('plugins:badges:edit', kwargs={
+            'organizer': self.request.event.organizer.slug,
+            'event': self.request.event.slug,
+            'layout': form.instance.pk
+        }))
 
     def form_invalid(self, form):
         messages.error(self.request, _('We could not save your changes. See below for details.'))
@@ -116,5 +128,55 @@ class LayoutDelete(EventPermissionRequiredMixin, DeleteView):
         })
 
 
-class LayoutEditorView(View):
-    pass
+class LayoutEditorView(BaseEditorView):
+    @cached_property
+    def layout(self):
+        try:
+            return self.request.event.badge_layouts.get(
+                id=self.kwargs['layout']
+            )
+        except BadgeLayout.DoesNotExist:
+            raise Http404(_("The requested badge layout does not exist."))
+
+    @property
+    def title(self):
+        return _('Badge layout: {}').format(self.layout)
+
+    def save_layout(self):
+        self.layout.layout = self.request.POST.get("data")
+        self.layout.save(update_fields=['layout'])
+
+    def get_default_background(self):
+        return static('pretixplugins/badges/badge_default_a6l.pdf')
+
+    def generate(self, op: OrderPosition, override_layout=None, override_background=None):
+        Renderer._register_fonts()
+
+        buffer = BytesIO()
+        if override_background:
+            bgf = default_storage.open(override_background.name, "rb")
+        elif isinstance(self.layout.background, File):
+            bgf = default_storage.open(self.layout.background.name, "rb")
+        else:
+            bgf = open(finders.find('pretixplugins/badges/badge_default_a6l.pdf'), "rb")
+        r = Renderer(
+            self.request.event,
+            override_layout or self.get_current_layout(),
+            bgf,
+        )
+        p = canvas.Canvas(buffer, pagesize=pagesizes.A4)
+        r.draw_page(p, op.order, op)
+        p.save()
+        outbuffer = r.render_background(buffer, 'Badge')
+        return 'badge.pdf', 'application/pdf', outbuffer.read()
+
+    def get_current_layout(self):
+        return json.loads(self.layout.layout)
+
+    def get_current_background(self):
+        return self.layout.background.url if self.layout.background else self.get_default_background()
+
+    def save_background(self, f: CachedFile):
+        if self.layout.background:
+            self.layout.background.delete()
+        self.layout.background.save('background.pdf', f.file)
